src/pyproject.toml
[tool.poetry]
name = "future-yasiro"
version = "0.1.0"
description = ""
authors = ["XDeviation <191602098@qq.com>"]
license = "AGPL-3.0 license"
readme = "README.md"
packages = [{include = "future_yasiro"}]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
sqlalchemy = "^1.4.45"
requests = "^2.28.1"
pika = "^1.3.1"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
